JK: added project BrickController2.Tools
