+++ BrickController2.sln
@@ -15,6 +15,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{02EA681E
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BrickController2.Tests", "BrickController2\BrickController2.Tests\BrickController2.Tests.csproj", "{7981CBBF-BB2F-4AF4-BEF3-3CF0C92DBB23}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{5AFB0DE9-536E-4C5D-831A-421ADC3EF3D9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BrickController2.Tools", "BrickController2\BrickController2.Tools\BrickController2.Tools.csproj", "{C127AB50-E104-2345-5239-13258E2B7474}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -103,12 +107,25 @@ Global
 		{7981CBBF-BB2F-4AF4-BEF3-3CF0C92DBB23}.Release|ARM64.Build.0 = Release|Any CPU
 		{7981CBBF-BB2F-4AF4-BEF3-3CF0C92DBB23}.Release|x64.ActiveCfg = Release|Any CPU
 		{7981CBBF-BB2F-4AF4-BEF3-3CF0C92DBB23}.Release|x64.Build.0 = Release|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Debug|x64.Build.0 = Debug|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Release|ARM64.Build.0 = Release|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Release|x64.ActiveCfg = Release|Any CPU
+		{C127AB50-E104-2345-5239-13258E2B7474}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{7981CBBF-BB2F-4AF4-BEF3-3CF0C92DBB23} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{C127AB50-E104-2345-5239-13258E2B7474} = {5AFB0DE9-536E-4C5D-831A-421ADC3EF3D9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {024A01F8-4022-4F9E-A006-AC2CABB04620}
