+++ BrickController2/BrickController2.Tools/BrickController2.Tools.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+	<TargetFramework>net9.0</TargetFramework>
+	<RootNamespace>BrickController2.Tools</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="BrickController2.Tests" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BrickController2\BrickController2.csproj" />
+  </ItemGroup>
+
+</Project>
