+++ BrickController2/BrickController2/Protocols/CryptTools.cs
@@ -74,69 +74,6 @@ public static class CryptTools
         return resultArrayLength;
     }
 
-    public static byte[] DecryptRfPayload(byte[] seed, int headerLength, int dataLength, int headerOffset, byte ctxValue1, byte ctxValue2, byte[] rfPayload)
-    {
-        const int checksumLength = 2;
-        int seedLength = seed.Length;
-        int resultArrayLength = headerLength + seedLength + dataLength + checksumLength;
-
-        int seedOffset = headerOffset + headerLength;
-        int dataOffset = seedOffset + seedLength;
-        int checksumOffset = dataOffset + dataLength;
-
-        int resultBufferLength = checksumOffset + checksumLength;
-
-        // Prepare buffer and copy payload into the correct offset
-        byte[] resultBuffer = new byte[resultBufferLength];
-        Buffer.BlockCopy(rfPayload, 0, resultBuffer, headerOffset, resultArrayLength);
-
-        // Reverse WhiteningEncode with ctxValue2 (same as encode, since it's XOR-based)
-        byte[] ctxArray2 = new byte[7];
-        WhiteningInit(ctxValue2, ctxArray2);
-        WhiteningEncode(resultBuffer, 0, resultBufferLength, ctxArray2);
-
-        // Reverse WhiteningEncode with ctxValue1 (same as encode, since it's XOR-based)
-        byte[] ctxArray1 = new byte[7];
-        WhiteningInit(ctxValue1, ctxArray1);
-        WhiteningEncode(resultBuffer, seedOffset, seedLength + dataLength + checksumLength, ctxArray1);
-
-        // Extract and invert header+seed
-        byte[] header = new byte[headerLength];
-        byte[] seedReversed = new byte[seedLength];
-        for (int i = 0; i < headerLength; i++)
-        {
-            header[i] = Invert8(resultBuffer[headerOffset + i]);
-        }
-
-        for (int i = 0; i < seedLength; i++)
-        {
-            seedReversed[i] = Invert8(resultBuffer[headerOffset + headerLength + i]);
-        }
-
-        // Reverse the seed array to get the original order
-        byte[] seedOriginal = new byte[seedLength];
-        for (int i = 0; i < seedLength; i++)
-        {
-            seedOriginal[i] = seedReversed[seedLength - 1 - i];
-        }
-
-        // Extract data
-        byte[] data = new byte[dataLength];
-        Buffer.BlockCopy(resultBuffer, dataOffset, data, 0, dataLength);
-
-        // Optionally, verify checksum
-        ushort expectedCrc = resultBuffer.GetUInt16(checksumOffset);
-        ushort actualCrc = CheckCRC16(seedOriginal, data);
-
-        if (expectedCrc != actualCrc)
-        {
-            throw new InvalidOperationException("CRC check failed. Data may be corrupted or parameters are incorrect.");
-        }
-
-        // Return the original data array
-        return data;
-    }
-
     /// <summary>
     /// inverts the bits of a given byte
     /// </summary>
